disassemble: extract TypeCache
