+++ strcs/annotations.py
@@ -29,7 +29,8 @@ from .decorator import (
     ConvertResponse,
     WrappedCreator,
 )
-from .disassemble.base import Type, TypeCache
+from .disassemble.base import Type
+from .disassemble.cache import TypeCache
 from .meta import Meta
 
 if tp.TYPE_CHECKING:
