+++ strcs/disassemble/base.py
@@ -23,6 +23,7 @@ from .score import Score
 if tp.TYPE_CHECKING:
     from ..annotations import AdjustableCreator, AdjustableMeta
     from ..decorator import ConvertFunction
+    from .cache import TypeCache
     from .type_tree import MRO
 
 
@@ -445,42 +446,3 @@ class Type(tp.Generic[T]):
     @memoized_property
     def checkable(self) -> type[InstanceCheck]:
         return create_checkable(self)
-
-
-class TypeCache(collections.abc.MutableMapping[object, "Type"]):
-    def __init__(self):
-        self.cache = {}
-
-    def key(self, o: object) -> tuple[type, object]:
-        return (type(o), o)
-
-    def __getitem__(self, k: object) -> Type:
-        return self.cache[self.key(k)]
-
-    def __setitem__(self, k: object, v: Type) -> None:
-        try:
-            hash(k)
-        except TypeError:
-            return
-        else:
-            self.cache[self.key(k)] = v
-
-    def __delitem__(self, k: object) -> None:
-        del self.cache[self.key(k)]
-
-    def __contains__(self, k: object) -> bool:
-        try:
-            hash(k)
-        except TypeError:
-            return False
-        else:
-            return self.key(k) in self.cache
-
-    def __iter__(self) -> tp.Iterator[object]:
-        return iter(self.cache)
-
-    def __len__(self) -> int:
-        return len(self.cache)
-
-    def clear(self) -> None:
-        self.cache.clear()
