+++ strcs/disassemble/type_tree.py
@@ -7,7 +7,8 @@ import attrs
 
 from ..memoized_property import memoized_property
 from ..standard import union_types
-from .base import Field, Type, TypeCache
+from .base import Field, Type
+from .cache import TypeCache
 
 
 class HasOrigBases(tp.Protocol):
